move charts regressions to regressions folder
