--- docs/data/visual-regression-tests/XAxisAnchorBaselineDefaults.tsx.preview
@@ -1,15 +1,0 @@
-<BarChart
-  xAxis={xAxes}
-  // Other props
-  width={600}
-  height={900}
-  dataset={usAirportPassengers}
-  series={[{ dataKey: '2022', label: '2022' }]}
-  hideLegend
-  yAxis={[
-    {
-      valueFormatter: (value: number) => `${(value / 1000).toLocaleString()}k`,
-      width: 60,
-    },
-  ]}
-/>
