+++ test/regressions/charts/BarBorderRadius.tsx
