+++ test/regressions/charts/XAxisTickLabelOverflow.tsx
