+++ test/regressions/charts/YAxisTickLabelOverflow.tsx
